requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0

plotly>=5.18

transformers>=4.41
torch>=2.0
sentence-transformers>=2.6
faiss-cpu>=1.7.4
smolagents>=0.2.0

python-dateutil>=2.8
